+++ client/src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Main from './components/main';
+import Main from './components/Main';
 import './App.css';
